фикс после ревью - удаление интерфейсов из слоя репозиториев, перенос в сервисный слой
