+++ inventory/internal/service/part/service.go
@@ -1,17 +1,16 @@
 package part
 
 import (
-	"github.com/alexis871aa/microservices-rocket-factory/inventory/internal/repository"
 	def "github.com/alexis871aa/microservices-rocket-factory/inventory/internal/service"
 )
 
 var _ def.PartService = (*service)(nil)
 
 type service struct {
-	partRepository repository.PartRepository
+	partRepository def.PartRepository
 }
 
-func NewService(partRepository repository.PartRepository) *service {
+func NewService(partRepository def.PartRepository) *service {
 	return &service{
 		partRepository: partRepository,
 	}
