+++ inventory/internal/service/service.go
@@ -10,3 +10,8 @@ type PartService interface {
 	GetPart(ctx context.Context, uuid string) (model.Part, error)
 	ListParts(ctx context.Context, filter model.PartsFilter) ([]model.Part, error)
 }
+
+type PartRepository interface {
+	GetPart(ctx context.Context, uuid string) (model.Part, error)
+	ListParts(ctx context.Context, filter model.PartsFilter) ([]model.Part, error)
+}
